requirements.txt
streamlit
google-generativeai
requests
python-dotenv
plotly
matplotlib
statsmodels
openpyxl